feat(index): Update more content on the homepage
Added <div> elements to contain more content for the
homepage including wrapping the <ul> in a <div>,
adding a <p> and an <img> elements. This is to add
more content on the page. Also added class attributes
to <div> elements to enable application of css styles
to the elements and the content they contain.